+++ client/src/pages/screens/cryptoCurrency/CryptoCurrency.jsx
@@ -371,7 +371,6 @@ const CryptoCurrencies = ({ globalLoading, user, updateGlobalValues }) => {
             </tr>
           );
         })}
-        {/* {actionButtons} */}
       </tbody>
     </Table>
   );